only authenticated users can access calender page, Alumni page and Gallery
