+++ src/app/calender/page.js
@@ -1,171 +1,40 @@
-'use client';
-import React, { useState, useRef, useEffect } from 'react';
-import DateTimePicker from './DateTimePicker';
-import './calendar.css';
-import FullCalendar from '@fullcalendar/react';
-import dayGridPlugin from '@fullcalendar/daygrid';
-import timeGridPlugin from '@fullcalendar/timegrid';
-import interactionPlugin from '@fullcalendar/interaction';
-import { eventSave, fetchEventsFromDB, auth, signInoutWithGoogle, issignedin } from "@/lib/firebase";
-import { Card } from "@chakra-ui/react";
-import { onAuthStateChanged } from "firebase/auth";
+'use client'
+import React, { useState, useEffect } from 'react';
+import { Flex, Heading, Button } from "@chakra-ui/react";
+import { checkIfUserSignedIn } from "@/lib/firebase";
+import Calendar from "./calenderpage";
 
-export default function CalendarComponent() {
-    const calendarRef = useRef(null);
-    const [selectedDate, setSelectedDate] = useState(new Date());
-    const [selectedTime, setSelectedTime] = useState('00:00');
-    const [showAddEventModal, setShowAddEventModal] = useState(false);
-    const [eventTitle, setEventTitle] = useState('');
-    const [eventDesc,setEventDesc]=useState('');
-    const [events, setEvents] = useState([]);
-    const [user, setUser] = useState(null);
+export default function CalenderPage (){
+  const [userSignedIn, setUserSignedIn] = useState(false);
 
-    useEffect(() => {
-        loadEventsFromFirebase();
-
-        const unsubscribe = onAuthStateChanged(auth, (user) => {
-            setUser(user ? user : null);
-        });
-
-        return () => unsubscribe();
-    }, []);
-
-    const loadEventsFromFirebase = async () => {
-        try {
-            const eventsData = await fetchEventsFromDB();
-            setEvents(eventsData);
-        } catch (error) {
-            console.error("Error loading events from Firebase:", error);
-        }
+  useEffect(() => {
+    const checkSignIn = async () => {
+      const user = await checkIfUserSignedIn();
+      if (user) {
+        console.log('User is signed in:', user);
+        setUserSignedIn(true);
+      } else {
+        console.log('No user signed in');
+        setUserSignedIn(false);
+      }
     };
+    checkSignIn();
+  }, []);
 
-    const handleClose = () => {
-        setShowAddEventModal(false);
-    };
+  const signInRedirect = () => {
+    window.location.href = '/login';
+  };
 
-    const handleEventTitleChange = (event) => {
-        setEventTitle(event.target.value);
-    };
-    const handleEventDescChange = (event) => {
-        setEventDesc(event.target.value);
-    };
-
-    const submit = (e) => {
-        e.preventDefault();
-        addEvent();
-        setEventTitle('');
-        setEventDesc('');
-        setSelectedDate(new Date());
-        setSelectedTime('00:00');
-    };
-
-    const handleDateClick = (arg) => {
-        if (user) {
-            setSelectedDate(arg.date);
-            setShowAddEventModal(true);
-        } else {
-            alert('Please log in to add events.');
-        }
-    };
-
-    const handleEventClick = (arg) => {
-        console.log('Event clicked:', arg.event);
-    };
-
-    const handleAddEvent = () => {
-        if (user) {
-            setShowAddEventModal(true);
-        } else {
-            alert('Please log in to add events.');
-        }
-    };
-
-    const addEvent = () => {
-        const newEvent = {
-            title: eventTitle,
-            decsription:eventDesc,
-            date: selectedDate.toISOString().split('T')[0],
-            time: selectedTime,
-            timestamp: new Date().getTime()
-        };
-        eventSave(newEvent);
-        setEvents([...events, newEvent]);
-        setShowAddEventModal(false);
-    };
-
-    const futureEvents = events.filter(event => new Date(event.date + 'T' + event.time) > new Date());
-
-    const eventList = events.map((event, index) => ({
-        title: event.title,
-        start: event.date + 'T' + event.time
-    }));
-
-    const handleEventContent = (arg) => {
-        return (
-            <div className="custom-event">
-                <div className="event-title">{arg.event.title}</div>
-                <div className="event-time">{arg.timeText}</div>
-            </div>
-        );
-    };
-
-    return (
-        <div style={{ paddingTop: "10vh", display: 'flex' }}>
-            <div className='bar'>
-                <Card className='card'>
-                    <FullCalendar
-                        ref={calendarRef}
-                        plugins={[dayGridPlugin, timeGridPlugin, interactionPlugin]}
-                        initialView='dayGridMonth'
-                        headerToolbar={{
-                            start: "today prev,next",
-                            center: "title",
-                            end: ""
-                        }}
-                        height={"100vh"}
-                        dateClick={handleDateClick}
-                        eventClick={handleEventClick}
-                        eventContent={handleEventContent}
-                        events={eventList}
-                    />
-                </Card>
-            </div>
-            <div className='sidebar'>
-                {/* <button onClick={signInoutWithGoogle} className='authButton'>
-                    {user ? 'Sign Out' : 'Sign In with Google'}
-                </button> */}
-                <button onClick={handleAddEvent} className='addbutton'>Add Event</button>
-                <h1 style={{ textAlign: 'center', fontWeight: 'bold', color: 'rgb(5, 5, 68)' }}>Upcoming Events</h1>
-
-                <ul className="event-list">
-                    {futureEvents.map((event, index) => (
-                        <li key={index} style={{ marginBottom: '10px', padding: '10px', backgroundColor: '#f7f7f7', borderRadius: '5px', boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)' }}>
-                            <span className="event-title" style={{ fontWeight: 'bold' ,marginLeft: '10px'}}>{event.title}</span>
-                            <span className="event-date" style={{ color: '#666', marginLeft: '10px' }}>{event.date}</span>
-                            <br></br>
-                            <span className="event-date" style={{ color: 'grey', marginLeft: '10px' }}>{event.decsription}</span>
-                            <span className="event-time" style={{ color: '#666', marginLeft: '10px' }}>{event.time}</span>
-                        </li>
-                    ))}
-                </ul>
-            </div>
-
-            {showAddEventModal && (
-                <div className="add-event-form">
-                    <p className='title' style={{ fontSize: 16 }}>{selectedDate.toDateString()}</p>
-                    <label>
-                        Event Title
-                        <input type="text" value={eventTitle} onChange={handleEventTitleChange} />
-                    </label>
-                    <label>
-                        Event Description
-                        <input type="text" value={eventDesc} onChange={handleEventDescChange} />
-                    </label>
-                    <DateTimePicker selectedDate={selectedDate} setSelectedDate={setSelectedDate} selectedTime={selectedTime} setSelectedTime={setSelectedTime} />
-                    <button type="submit" onClick={submit}>Submit</button>
-                    <button type="button" onClick={handleClose}>Close</button>
-                </div>
-            )}
-        </div>
-    );
-}
+  return (
+    <>
+      {!userSignedIn ? (
+        <Flex justifyContent="center" alignItems="center" h="60vh" flexDirection="column">
+          <Heading fontSize="4xl" color="gray.600" textAlign="center">You need to sign in first</Heading>
+          <Button colorScheme="blue" mt={4} onClick={signInRedirect}>Sign In</Button>
+        </Flex>
+      ) : (
+        <Calendar/>
+      )}
+    </>
+  );
+};
